Refactor page structure and add HomeClient component for employee data display
