+++ src/types/employee.ts
@@ -1,0 +1,7 @@
+export type Employee = {
+  id: number;
+  firstName: string;
+  lastName: string;
+  location: string;
+  birthDate: string;
+};
